Fix problem with test script on Julia 0.2 (conversions-in-constructor issue)

+++ test/require.jl
@@ -5,7 +5,7 @@ using HDF5, JLD
 include("JLDTest.jl")
 
 function create()
-    x = JLDTest(5)
+    x = JLDTest(int16(5))  # int16 makes this work on 0.2
     jldopen("require.jld", "w") do file
         addrequire(file, joinpath(Pkg.dir(), "HDF5", "test", "JLDTest.jl"))
         write(file, "x", x, rootmodule="JLDTemp")
